Use int instead of long
It is unlikely for long to ever be reached in one session.
There can only be max int waypoints at a time. For this to run out of bounds, a use must create integer waypoints and delete some and replace them with the same amount.

+++ src/main/java/de/ecconia/scrapmechanicmapper/GPSWindow.java
@@ -353,8 +353,8 @@ public class GPSWindow extends JFrame
 	{
 		private float zoomFactor = 1f;
 		private boolean isMouseIn;
-		private long highlightedWaypointID;
-		private long selectedWaypointID;
+		private int highlightedWaypointID;
+		private int selectedWaypointID;
 		
 		public DrawPane()
 		{
